pyproject.toml
[tool.poetry]
name = "eurosat-landuse-classifier"
version = "0.1.0"
description = "Land use / land cover classification on EuroSAT RGB satellite images"
authors = ["Dmitriy Galkin"]
readme = "README.md"
packages = [{ include = "eurosat_landuse" }]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
torch = ">=2.2.0"
torchvision = ">=0.17.0"
pytorch-lightning = ">=2.2.0"
torchmetrics = ">=1.3.0"
hydra-core = ">=1.3.2"
omegaconf = ">=2.3.0"
mlflow = ">=2.10.0"
fire = ">=0.6.0"
pillow = ">=10.0.0"
matplotlib = ">=3.8.0"
numpy = ">=1.26.0"
tqdm = ">=4.66.0"
dvc = ">=3.40.0"
requests = ">=2.31.0"
pandas = ">=2.0.0"

[tool.poetry.group.dev.dependencies]
pre-commit = ">=3.6.0"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100

[tool.flake8]
max-line-length = 100
extend-ignore = ["E203", "W503"]
exclude = [".git", ".venv", "mlruns", ".dvc", "data", "artifacts"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
